+++ src/components/header/components/nav-menu/styles.js
@@ -1,10 +1,10 @@
-@import '~styles/settings.scss';
+import styled from '@emotion/styled';
 
-.c-nav-menu {
+export const NavMenuWrapper = styled.ul`
   display: flex;
   align-items: center;
   height: 100%;
   justify-content: flex-start;
   flex: 1;
   z-index: 1;
-}
+`;
