--- src/components/header/styles.scss
@@ -1,151 +1,0 @@
-@import '~styles/index.scss';
-
-$header-height: 56px;
-
-.c-header {
-  height: $header-height;
-  background: $white;
-  position: relative;
-  width: 100%;
-  border-bottom: solid 1px $border;
-
-  &.relative {
-    height: auto;
-    min-height: $header-height;
-  }
-
-  .logo {
-    position: absolute;
-    top: 0;
-    z-index: 2;
-    width: 76px;
-    height: 76px;
-  }
-
-  .nav {
-    display: flex;
-    position: relative;
-    justify-content: flex-end;
-    height: $header-height;
-
-    padding-left: 0;
-
-    @media screen and (min-width: $screen-m-bg) {
-      justify-content: flex-start;
-      padding-left: 80px;
-    }
-
-    @media screen and (min-width: 850px) {
-      padding-left: 90px;
-    }
-
-    &.mobile {
-      padding-left: 0;
-    }
-  }
-
-  .nav-desktop {
-    display: flex;
-    justify-content: flex-end;
-    position: relative;
-    height: $header-height;
-    width: 100%;
-
-    &.show-menu {
-      justify-content: space-between;
-    }
-  }
-
-  .nav-mobile {
-    display: flex;
-
-    @media screen and (min-width: 860px) {
-      display: none;
-    }
-  }
-
-  .nav-item {
-    height: 100%;
-    position: relative;
-
-    .nested {
-      padding: 0;
-      margin: 0;
-    }
-
-    > a,
-    > button,
-    > div,
-    .nav-link {
-      height: 100%;
-      display: flex;
-      align-items: center;
-      color: $slate;
-      font-size: 14px;
-      text-transform: uppercase;
-      text-decoration: none;
-      padding: 0 10px;
-      margin: 0 5px;
-      position: relative;
-
-      svg {
-        fill: $slate;
-        margin-top: -2px;
-        transition: all 150ms ease-out;
-      }
-
-      &:hover {
-        color: darken($slate, 30%);
-
-        svg {
-          fill: darken($slate, 30%);
-        }
-      }
-
-      &:focus {
-        outline: none;
-      }
-
-      &::after {
-        content: '';
-        height: 5px;
-        position: absolute;
-        bottom: 0;
-        right: 0;
-        left: 0;
-        background-color: transparent;
-      }
-
-      &.active {
-        &::after {
-          background-color: $green;
-        }
-      }
-    }
-  }
-
-  .nav-more > .nav-link {
-    padding-right: 0;
-    margin-right: 0;
-  }
-
-  .icon-submenu {
-    margin-left: 5px;
-  }
-
-  .icon-close {
-    width: 15px;
-    height: 10px;
-  }
-
-  .icon-menu {
-    width: 15px;
-    height: 15px;
-  }
-
-  .my-gfw-icon {
-    width: 15px;
-    height: 15px;
-    min-width: 15px;
-  }
-}
